CarRentalProject/src/test/java/com/carrental/service/HomeServiceTest.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/UnitTests/JUnit5TestClass.java to edit this template
 */
package com.carrental.service;

import com.carrental.dto.HomeDto;

import java.time.LocalDate;

import static org.assertj.core.api.Assertions.assertThat;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.Test;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.ActiveProfiles;

/**
 *
 * @author tomeku
 */
@SpringBootTest
@ActiveProfiles("test")
public class HomeServiceTest {
    
    @Autowired
    private HomeService homeService;
        
    public HomeServiceTest() {
    }
    
    @BeforeAll
    public static void setUpClass() {
    }
    
    @AfterAll
    public static void tearDownClass() {
    }
    
    @BeforeEach
    public void setUp() {
    }
    
    @AfterEach
    public void tearDown() {
    }

    /**
     * Test of getHome method, of class HomeService.
     */
    @Test
    public void testGetHome() {
        System.out.println("getHome");
        
        HomeDto homeDto = homeService.getHome();
        
        assertThat(homeDto.getListOfCities().size()).isEqualTo(4);
        assertThat(homeDto.getDate()).isEqualTo(LocalDate.now().toString());
        
    }
    
}
